+++ src/main/java/br/com/lcv/usuario/UsuarioRepository.java
@@ -11,6 +11,6 @@ import java.util.Optional;
 @Repository
 public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
 
-    Optional<Usuario> findUsuarioByTelegramId(int telegramId);
+    Optional<Usuario> findUsuarioByTelegramId(long telegramId);
 
 }
